feat: swap to pre-built binaries
Due to building this image for arm64 we need to use pre-built binaries
as building veilid from within QEMU is far too slow and exhausted my
github minutes. Might look into build the binaries and then creating the
container for full ownership of the supply line but for now will rely on
upstream rpm repo.